Move zoom-out logic to selector instead of action
